Heavily refactored and renamed intro qmodule, now userinfo module
